Add SQL-standard multiple linear self-references in WITH RECURSIVE
The recursive term of a WITH RECURSIVE query must contain exactly one
self-reference. Multiple such self-references are only allowed if the WITH query
features multiple recursive terms, all of which are joined by UNION [ALL].

Without using explicit parentheses to change the precedence of the parsing,
a left-deep UNION [ALL] tree is created. This allows for multiple non-recursive
terms, and exactly one recursive term. To allow multiple non-recursive terms,
as well as multiple recursive terms, a different grouping of terms is required.
To fix this, tree rotation is added to checkWellFormedRecursion.

Example:
A, B, and C are arbitrary SelectStmt nodes and can be deeper nested UNION nodes.

A is a non-recursive term in the WITH RECURSIVE query. B, and C both contain a
recursive self-reference. The planner expects the top UNION node to contain
the non-recursive term in the larg, and the recursive term in the rarg.
Therefore, the tree shape on the left is invalid and would result in an error
message at the parsing stage. However, by rotating the tree to the right, this
problem can be solved so that the valid tree shape on the right side is created.

      UNION   --->   UNION
     /     \        /     \
   UNION    C      A    UNION
  /     \              /     \
 A       B            B       C

Effectively this re-parenthesizes the expression:
(A UNION B) UNION C ---> A UNION (B UNION C)