Improve jitting performance by not emitting the LLVMPassManagerBuilderUseInlinerWithThreshold pass.
This pass contains some very expensive parts which in experiments so
far do not gain us much compared to the runtime we gain afterwards
when running queries. Instead now emit the simpler inliner passes
also for PGJIT_OPT3.

To monitor how many modules are now created an extra statistic is
added so we can test effectively how much performance is
gained / lost because of this.