relcache not invalidated when ADD PRIMARY KEY USING INDEX
Hi,

When reviewing some replica identity related patches, I found that when adding
primary key using an existing unique index on not null columns, the
target table's relcache won't be invalidated.

This would cause an error When REPLICA IDENTITY is default and we are
UPDATE/DELETE a published table , because we will refer to
RelationData::rd_pkindex to check if the UPDATE or DELETE can be safety
executed in this case.

---reproduction steps
CREATE TABLE test(a int not null, b int);
CREATE UNIQUE INDEX a ON test(a);
CREATE PUBLICATION PUB for TABLE test;
UPDATE test SET a = 2;
	ERROR:  cannot update table "test" because it does not have a replica identity and publishes updates
	HINT:  To enable updating the table, set REPLICA IDENTITY using ALTER TABLE.

alter table test add primary key using index a;
UPDATE test SET a = 2;
	ERROR:  cannot update table "test" because it does not have a replica identity and publishes updates
	HINT:  To enable updating the table, set REPLICA IDENTITY using ALTER TABLE.