cirrus: build docs as a separate task..
Always run doc build; to allow them to be shown in cfbot.
A failure in documentation should be shown even if Linux fails check-world.

This'll automatically show up as a separate "column" on cfbot.

Also, in the future, this will hopefully upload each patch's changed HTML docs
as an artifact, for easy review.

XX-os-only: html