Re-bin segment when dsa memory is freed.
Extract re-bin segment into a separate function called rebin_segment,
call it to ajust the segment to appropriate bin when dsa memory is
freed. Otherwise, when allocating memory, due to the segment with
enough contiguous pages is in a smaller bin, a suitable segment
may not be found to allocate memory.