WIP: cirrus/windows: cache windows source and build dirs..
..to enable incremental re-build.

I've had minimal success using any of the compiler caches (clcache, cclache,
buildcache, ccache, sccache).  This can reduce build time from ~3.5min to
~2.5min.  But since the windows VM takes ~4min to start, reducing build time by
only ~1min is an improvement of little significance.  The goal is to reduce the
build time to closer to ~15sec, and the entire CI instance closer to the rest
of the OS: (10min rather than 25min).

ci-os-only: windows

https://cirrus-ci.com/task/6479098905100288
https://cirrus-ci.com/task/6717456000483328