WIP: do not install testclient nor uri-regress
Since the uri-regress and testclient tools are in a subdir of the testdir, the
PATH needs to be additionally manipulated.  This allows not installing the
tools, which are only useful for running tap tests.

NO: This reverts commit a17fd67d2f2861ae0ce00d1aeefdf2facc47cd5e.

ci-os-only: windows-msvc